strip platforms
