+++ src/main/scala/io/viash/executors/NextflowExecutor.scala
@@ -17,8 +17,20 @@
 
 package io.viash.executors
 
-import io.viash.schemas._
 import io.viash.config.Config
+import io.viash.functionality._
+import io.viash.functionality.resources._
+import io.viash.functionality.arguments._
+import io.viash.helpers.{Docker, Bash, DockerImageInfo, Helper}
+import io.viash.helpers.circe._
+import io.viash.platforms.nextflow._
+import io.circe.syntax._
+import io.circe.{Printer => JsonPrinter, Json, JsonObject}
+import shapeless.syntax.singleton
+import io.viash.schemas._
+import io.viash.helpers.Escaper
+import io.viash.executors.{Executor, ExecutorResources}
+import io.viash.engines.DockerEngine
 
 final case class NextflowExecutor(
   @description("Name of the executor. As with all executors, you can give an executor a different name. By specifying `id: foo`, you can target this executor (only) by specifying `...` in any of the Viash commands.")
@@ -26,21 +38,303 @@ final case class NextflowExecutor(
   @default("nextflow")
   id: String = "nextflow",
 
-  `type`: String = "nextflow"
+  `type`: String = "nextflow",
+
+  // nxf params
+  @description(
+    """@[Directives](nextflow_directives) are optional settings that affect the execution of the process. These mostly match up with the Nextflow counterparts.  
+      |""".stripMargin)
+  @example(
+    """directives:
+      |  container: rocker/r-ver:4.1
+      |  label: highcpu
+      |  cpus: 4
+      |  memory: 16 GB""".stripMargin,
+      "yaml")
+  @default("Empty")
+  directives: NextflowDirectives = NextflowDirectives(),
+
+  @description(
+    """@[Automated processing flags](nextflow_auto) which can be toggled on or off:
+      |
+      || Flag | Description | Default |
+      ||---|---------|----|
+      || `simplifyInput` | If `true`, an input tuple only containing only a single File (e.g. `["foo", file("in.h5ad")]`) is automatically transformed to a map (i.e. `["foo", [ input: file("in.h5ad") ] ]`). | `true` |
+      || `simplifyOutput` | If `true`, an output tuple containing a map with a File (e.g. `["foo", [ output: file("out.h5ad") ] ]`) is automatically transformed to a map (i.e. `["foo", file("out.h5ad")]`). | `true` |
+      || `transcript` | If `true`, the module's transcripts from `work/` are automatically published to `params.transcriptDir`. If not defined, `params.publishDir + "/_transcripts"` will be used. Will throw an error if neither are defined. | `false` |
+      || `publish` | If `true`, the module's outputs are automatically published to `params.publishDir`.  Will throw an error if `params.publishDir` is not defined. | `false` |
+      |
+      |""".stripMargin)
+  @example(
+    """auto:
+      |  publish: true""".stripMargin,
+      "yaml")
+  @default(
+    """simplifyInput: true
+      |simplifyOutput: true
+      |transcript: false
+      |publish: false
+      |""".stripMargin)
+  auto: NextflowAuto = NextflowAuto(),
+
+  @description("Allows tweaking how the @[Nextflow Config](nextflow_config) file is generated.")
+  @since("Viash 0.7.4")
+  @default("A series of default labels to specify memory and cpu constraints")
+  config: NextflowConfig = NextflowConfig(),
+
+  @description("Whether or not to print debug messages.")
+  @default("False")
+  debug: Boolean = false,
+
+  // TODO: solve differently
+  @description("Specifies the Docker platform id to be used to run Nextflow.")
+  @default("docker")
+  container: String = "docker"
 ) extends Executor {
+  def escapeSingleQuotedString(txt: String): String = {
+    Escaper(txt, slash = true, singleQuote = true, newline = true)
+  }
+
   def generateExecutor(config: Config, testing: Boolean): ExecutorResources = {
-    val engines = config.getEngines
-    
-    // todo: do something with containers
+    val condir = containerDirective(config)
 
-    // todo: generate mainscript
-    val mainScript = None
-    val additionalResources = Nil
-
-    // return output
-    ExecutorResources(
-      mainScript = mainScript,
-      additionalResources = additionalResources
+    // create main.nf file
+    val mainFile = PlainFile(
+      dest = Some("main.nf"),
+      text = Some(renderMainNf(config, condir))
     )
+    val nextflowConfigFile = PlainFile(
+      dest = Some("nextflow.config"),
+      text = Some(renderNextflowConfig(config.functionality, condir))
+    )
+
+    // remove main
+    val otherResources = config.functionality.additionalResources
+
+    ExecutorResources(
+      None,
+      mainFile :: nextflowConfigFile :: otherResources
+    )
+  }
+
+  def containerDirective(config: Config): Option[DockerImageInfo] = {
+    val plat = config.getEngines.find(p => p.id == container)
+    plat match {
+      case Some(p: DockerEngine) => 
+        Some(Docker.getImageInfo(
+          functionality = Some(config.functionality),
+          engineId = Some(p.id),
+          registry = p.target_registry,
+          organization = p.target_organization,
+          name = p.target_image,
+          tag = p.target_tag.map(_.toString),
+          namespaceSeparator = p.namespace_separator
+        ))
+      case Some(_) => 
+        throw new RuntimeException(s"NextflowPlatform 'container' variable: Platform $container is not a Docker Platform")
+      case None => None
+      case _ => ???
+    }
+  }
+
+  def renderNextflowConfig(functionality: Functionality, containerDirective: Option[DockerImageInfo]): String = {
+    val versStr = functionality.version.map(ver => s"\n  version = '$ver'").getOrElse("")
+
+    val descStr = functionality.description.map{des => 
+      val escDes = escapeSingleQuotedString(des)
+      s"\n  description = '$escDes'"
+    }.getOrElse("")
+
+    val authStr = 
+      if (functionality.authors.isEmpty) {
+        "" 
+      } else {
+        val escAut = escapeSingleQuotedString(functionality.authors.map(_.name).mkString(", "))
+        s"\n  author = '$escAut'"
+      }
+
+    // TODO: define profiles
+    val profileStr = 
+      if (containerDirective.isDefined || functionality.mainScript.map(_.`type`) == Some(NextflowScript.`type`)) {
+        "\n\n" + NextflowHelper.profilesHelper
+      } else {
+        ""
+      }
+
+    val processLabels = config.labels.map{ case (k, v) => s"withLabel: $k { $v }"}
+    val inlineScript = config.script.toList
+
+    s"""manifest {
+    |  name = '${functionality.name}'
+    |  mainScript = 'main.nf'
+    |  nextflowVersion = '!>=20.12.1-edge'$versStr$descStr$authStr
+    |}$profileStr
+    |
+    |process{
+    |  ${processLabels.mkString("\n  ")}
+    |}
+    |
+    |${inlineScript.mkString("\n")}
+    |""".stripMargin
+  }
+
+  // interpreted from BashWrapper
+  def renderMainNf(config: Config, containerDirective: Option[DockerImageInfo]): String = {
+    val functionality = config.functionality
+    
+    /************************* HEADER *************************/
+    val header = Helper.generateScriptHeader(functionality)
+      .map(h => Escaper(h, newline = true))
+      .mkString("// ", "\n// ", "")
+    
+    /************************* JSONS *************************/
+    // override container
+    val directivesToJson = directives.copy(
+      // if a docker platform is defined but the directives.container isn't, use the image of the dockerplatform as default
+      container = directives.container orElse containerDirective.map(cd => Left(cd.toMap)),
+      // is memory requirements are defined but directives.memory isn't, use that instead
+      memory = directives.memory orElse functionality.requirements.memoryAsBytes.map(_.toString + " B"),
+      // is cpu requirements are defined but directives.cpus isn't, use that instead
+      cpus = directives.cpus orElse functionality.requirements.cpus.map(np => Left(np))
+    )
+    val jsonPrinter = JsonPrinter.spaces2.copy(dropNullValues = true)
+    val dirJson = directivesToJson.asJson.dropEmptyRecursively
+    val dirJson2 = if (dirJson.isNull) Json.obj() else dirJson
+    val funJson = config.asJson.dropEmptyRecursively
+    val funJsonStr = jsonPrinter.print(funJson)
+      .replace("\\\\", "\\\\\\\\")
+      .replace("\\\"", "\\\\\"")
+      .replace("'''", "\\'\\'\\'")
+      .grouped(65000) // JVM has a maximum string limit of 65535
+      .toList         // see https://stackoverflow.com/a/6856773
+      .mkString("'''", "''' + '''", "'''")
+    val autoJson = auto.asJson.dropEmptyRecursively
+
+    /************************* MAIN.NF *************************/
+    functionality.mainScript match {
+      // if mainResource is empty (shouldn't be the case)
+      case None => throw new RuntimeException("there should be a main script here")
+
+      // if mainResource is simply an executable
+      case Some(e: Executable) => //" " + e.path.get + " $VIASH_EXECUTABLE_ARGS"
+        throw new NotImplementedError(
+          "Running executables through a NextflowPlatform is not (yet) implemented. " +
+            "Create a support ticket to request this functionality if necessary."
+        )
+      
+      case Some(res: NextflowScript) =>
+        res.readWithInjection(Map.empty, config).get
+        // TODO: change this such that it contains helper code instead of
+        // importing non-existent helper files
+
+      // if mainResource is a script
+      case Some(res) =>
+        // todo: also include the bashwrapper checks
+        val argsAndMeta = functionality.getArgumentLikesGroupedByDest(
+          includeMeta = true,
+          filterInputs = true
+        )
+        val code = res.readWithInjection(argsAndMeta, config).get
+        val escapedCode = Bash.escapeString(code, allowUnescape = true)
+          .replace("\\", "\\\\")
+          .replace("'''", "\\'\\'\\'")
+
+        // IMPORTANT! difference between code below and BashWrapper:
+        // script is stored as `.viash_script.sh`.
+        val scriptPath = "$tempscript"
+
+        val executionCode = 
+          s"""set -e
+            |tempscript=".viash_script.sh"
+            |cat > "$scriptPath" << VIASHMAIN
+            |$escapedCode
+            |VIASHMAIN
+            |${res.command(scriptPath)}
+            |""".stripMargin
+        
+        s"""$header
+          |
+          |nextflow.enable.dsl=2
+          |
+          |// initialise slurper
+          |def jsonSlurper = new groovy.json.JsonSlurper()
+          |
+          |// DEFINE CUSTOM CODE
+          |
+          |// functionality metadata
+          |thisConfig = processConfig(jsonSlurper.parseText($funJsonStr))
+          |
+          |thisScript = '''$executionCode'''
+          |
+          |thisDefaultProcessArgs = [
+          |  // key to be used to trace the process and determine output names
+          |  key: thisConfig.functionality.name,
+          |  // fixed arguments to be passed to script
+          |  args: [:],
+          |  // default directives
+          |  directives: jsonSlurper.parseText('''${jsonPrinter.print(dirJson2)}'''),
+          |  // auto settings
+          |  auto: jsonSlurper.parseText('''${jsonPrinter.print(autoJson)}'''),
+          |
+          |  // Apply a map over the incoming tuple
+          |  // Example: `{ tup -> [ tup[0], [input: tup[1].output] ] + tup.drop(2) }`
+          |  map: null,
+          |
+          |  // Apply a map over the ID element of a tuple (i.e. the first element)
+          |  // Example: `{ id -> id + "_foo" }`
+          |  mapId: null,
+          |
+          |  // Apply a map over the data element of a tuple (i.e. the second element)
+          |  // Example: `{ data -> [ input: data.output ] }`
+          |  mapData: null,
+          |
+          |  // Apply a map over the passthrough elements of a tuple (i.e. the tuple excl. the first two elements)
+          |  // Example: `{ pt -> pt.drop(1) }`
+          |  mapPassthrough: null,
+          |
+          |  // Filter the channel
+          |  // Example: `{ tup -> tup[0] == "foo" }`
+          |  filter: null,
+          |
+          |  // Rename keys in the data field of the tuple (i.e. the second element)
+          |  // Will likely be deprecated in favour of `fromState`.
+          |  // Example: `[ "new_key": "old_key" ]`
+          |  renameKeys: null,
+          |
+          |  // Fetch data from the state and pass it to the module without altering the current state.
+          |  // 
+          |  // `fromState` should be `null`, `List[String]`, `Map[String, String]` or a function. 
+          |  // 
+          |  // - If it is `null`, the state will be passed to the module as is.
+          |  // - If it is a `List[String]`, the data will be the values of the state at the given keys.
+          |  // - If it is a `Map[String, String]`, the data will be the values of the state at the given keys, with the keys renamed according to the map.
+          |  // - If it is a function, the tuple (`[id, state]`) in the channel will be passed to the function, and the result will be used as the data.
+          |  // 
+          |  // Example: `{ id, state -> [input: state.fastq_file] }`
+          |  // Default: `null`
+          |  fromState: null,
+          |
+          |  // Determine how the state should be updated after the module has been run.
+          |  // 
+          |  // `toState` should be `null`, `List[String]`, `Map[String, String]` or a function.
+          |  // 
+          |  // - If it is `null`, the state will be replaced with the output of the module.
+          |  // - If it is a `List[String]`, the state will be updated with the values of the data at the given keys.
+          |  // - If it is a `Map[String, String]`, the state will be updated with the values of the data at the given keys, with the keys renamed according to the map.
+          |  // - If it is a function, a tuple (`[id, output, state]`) will be passed to the function, and the result will be used as the new state.
+          |  //
+          |  // Example: `{ id, output, state -> state + [counts: state.output] }`
+          |  // Default: `{ id, output, state -> output }`
+          |  toState: null,
+          |
+          |  // Whether or not to print debug messages
+          |  // Default: `$debug`
+          |  debug: $debug
+          |]
+          |
+          |// END CUSTOM CODE""".stripMargin + 
+          "\n\n" + NextflowHelper.workflowHelper + 
+          "\n\n" + NextflowHelper.vdsl3Helper
+    }
   }
 }
